LayoutTests/accessibility/gtk/replaced-objects-in-anonymous-blocks.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN"> 
<html> 
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../resources/accessibility-helper.js"></script>
</head> 
<body> 
 
<p>Paragraph</p>
<button>button</button>
<p></p>
<iframe src="data:text/html,<body>iframe</body>"></iframe>
<p></p>
<textarea>text area</textarea>
<div>
Div
<p></p>
<button>button</button>
</div>
<h3>
Heading
<p></p>
<button>button</button>
<p></p>
<img src="foo" alt="I should be an image" />
</h3>

<div id="stopElement">End of test</div>

<pre id="tree"></pre>
 
<p id="description"></p> 
<div id="console"></div> 
 
<script>
    description("This tests that the text associated with replaced objects in an anonymous block is correct.");

    if (window.accessibilityController) {
        // Build the accessibility tree up until 'End of test' is encountered.
        window.stopElement = accessibilityController.accessibleElementById("stopElement");

        document.body.focus();
        dumpAccessibilityTree(accessibilityController.focusedElement, stopElement, 0);
    }

</script>
<script src="../../resources/js-test-post.js"></script>
</body> 
</html> 
